chore(keychain-sdk): move Encrypt to internal pkg
