+++ keychain-sdk/internal/enc/enc.go
@@ -1,4 +1,4 @@
-package keychain
+package enc
 
 import (
 	"crypto/ecdsa"
