+++ keychain-sdk/internal/enc/enc_test.go
@@ -1,4 +1,4 @@
-package keychain
+package enc
 
 import (
 	"testing"
